factored out electro 2d only functions
