--- backend/startup.py
@@ -1,10 +1,0 @@
-from fastapi import APIRouter
-
-router = APIRouter(
-    tags=['Startup']
-)
-
-
-@router.get("/")
-async def hello_world():
-    return dict({'message': 'Ready to go'})
